Move SMWT components to visualization folder
