+++ src/components/researcher/taskvisualizations/SmwtChartVisualization.vue
