+++ src/components/researcher/taskvisualizations/SmwtMapVisualization.vue
